+++ modules/platforms/cpp/ignite/odbc/string_utils_test.cpp
@@ -109,7 +109,7 @@ TEST_F(string_utils_test, split_basic)
 
         ASSERT_EQ(exp.size(), res.size());
 
-        for (int i = 0; i < exp.size(); ++i) {
+        for (size_t i = 0; i < exp.size(); ++i) {
             EXPECT_EQ(exp[i],res[i]) << "Vectors differ at index " << i;
         }
     };
